+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :bug:`8120` Ethereum staking daily stats should be available again.
 * :bug:`-` Fix the issue where the popup notification breaks the experience when the notification sidebar is already open.
 * :bug:`-` Fixed an issue with the assets updates where conflicts couldn't be resolved for multiple changes in a single asset.
 * :bug:`-` Stakedao bribe claims older than January 2023 will now also be properly decoded.